cli: split session command and cover environments
